[PATRO23-43] move Export/ImportResult to shared.infrastructure
