+++ src/modules/budget/application/Budget/ImportingBudgets/GetImportResult.cs
@@ -1,3 +1,5 @@
+using Intive.Patronage2023.Shared.Infrastructure.Import;
+
 namespace Intive.Patronage2023.Modules.Budget.Application.Budget.ImportingBudgets;
 
 /// <summary>
