+++ src/shared/infrastructure/Import/ImportResult.cs
@@ -1,4 +1,4 @@
-namespace Intive.Patronage2023.Modules.Budget.Application.Budget.ImportingBudgets;
+namespace Intive.Patronage2023.Shared.Infrastructure.Import;
 
 /// <summary>
 /// Class ImportResult representing the result of an import operation, containing any encountered error messages and the URI used for import.
